requirements.txt
python-chess>=1.999
streamlit>=1.50.0
pytest>=8.0.0